+++ src/main/java/com/ll/tourdemonde/reservation/repository/ReservationOptionRepository.java
@@ -1,6 +1,6 @@
-package com.ll.tourdemonde.reservation.reservation.repository;
+package com.ll.tourdemonde.reservation.repository;
 
-import com.ll.tourdemonde.reservation.reservation.entity.ReservationOption;
+import com.ll.tourdemonde.reservation.entity.ReservationOption;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ReservationOptionRepository extends JpaRepository<ReservationOption, Long> {
