+++ src/test/java/com/ll/tourdemonde/reservation/reservation/ReservationControllerTest.java
@@ -5,11 +5,11 @@ import com.ll.tourdemonde.place.dto.PlaceReqDto;
 import com.ll.tourdemonde.place.dto.PlaceReqDtoList;
 import com.ll.tourdemonde.place.entity.Place;
 import com.ll.tourdemonde.place.service.PlaceService;
-import com.ll.tourdemonde.reservation.reservation.controller.ReservationController;
-import com.ll.tourdemonde.reservation.reservation.entity.Reservation;
-import com.ll.tourdemonde.reservation.reservation.entity.ReservationOption;
-import com.ll.tourdemonde.reservation.reservation.entity.ReservationType;
-import com.ll.tourdemonde.reservation.reservation.service.ReservationService;
+import com.ll.tourdemonde.reservation.controller.ReservationController;
+import com.ll.tourdemonde.reservation.entity.Reservation;
+import com.ll.tourdemonde.reservation.entity.ReservationOption;
+import com.ll.tourdemonde.reservation.entity.ReservationType;
+import com.ll.tourdemonde.reservation.service.ReservationService;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.MethodOrderer;
 import org.junit.jupiter.api.Test;
